+++ pystac/stac_object.py
@@ -158,8 +158,8 @@ class STACObject(ABC):
             links have absolute (as opposed to relative) HREFs.
         """
         self_link = self.get_single_link(pystac.RelType.SELF)
-        if self_link:
-            return cast(str, self_link.target)
+        if self_link and self_link.has_target_href():
+            return self_link.get_target_str()
         else:
             return None